+++ neoforge/src/main/java/de/cech12/bucketlib/api/crafting/EmptyIngredient.java
@@ -67,7 +67,9 @@ public class EmptyIngredient extends Ingredient {
     public ItemStack[] getItems() {
         if (this.matchingStacks == null) {
             ArrayList<ItemStack> stacks = new ArrayList<>();
-            stacks.add(new ItemStack(Items.BUCKET));
+            if (this.item == null && this.tag == null) {
+                stacks.add(new ItemStack(Items.BUCKET));
+            }
             BucketLibMod.getRegisteredBuckets().forEach(universalBucketItem -> {
                 ItemStack universalBucketItemStack = new ItemStack(universalBucketItem);
                 if (this.item != null && universalBucketItem == this.item